Created modular texture system
WARNING: I use Windows and updated the Makefile ( see -lstdc++fs under Windows flags) in order to build.  Linux builds might need to have an updated Makefile, but I am unable to debug for Linux.
The numbering system is only for semantic use, it does not currently relate to in-game indexes unless every single texture slot is filled.
There are 256 available texture slots.  Attempting to include additional textures may result in an error.  If there are fewer than 256 textures, the remaining ones will be filled in black (except for texture 6)