Multi-thread Makefile for Linux
Separate sources from objects and create dependencies to allow for
multi-threading with `-j`.

Objects are built separately in `obj/`, which is created if needed, then
linked together.

This prevents unnecessary re-compilation of `.cpp` files but it will not detect
changes in header files. This could be fixed with a wild-card
dependency for `.h` files.